Splits PostGrid.jsx component
